# src/main/java/com/egguard/egguardbackend/egg/request/PickEggsRequest.java
package com.egguard.egguardbackend.egg.request;

import lombok.*;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class PickEggsRequest {
    // Optional filter: mark eggs picked before this timestamp
    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
    private LocalDateTime before;
}